- Fixes #1597 - Deletion of Tenant should also delete its modeshape workspace - Refactored IModelRepository. Moved methods not related to model   manipulation to IRepositoryManager - IRepositoryManager's responsibility now anything regarding the   repository as a whole (backup, restore, create/delete workspace)
